Code Refactoring
Code Abstraction. Breaking down our code into smaller chunks to make it easier to work on features and to prevent accumulating errors in single file.
