--- AppDevelopment/complete_recorder/lib/App Pages/recording_button.dart
@@ -1,158 +1,0 @@
-// TODO Implement this library.
-import 'dart:io';
-import 'package:flutter/material.dart';
-import 'package:flutter_audio_recorder/flutter_audio_recorder.dart';
-import 'package:path_provider/path_provider.dart';
-
-class RecordButton extends StatefulWidget {
-  final Function onSaved;
-
-  const RecordButton({Key key, @required this.onSaved}) : super(key: key);
-  @override
-  _RecordButtonState createState() => _RecordButtonState();
-}
-
-//checks on recording state
-enum RecordingState {
-  /// Recording not initialized
-  UnSet,
-
-  /// Ready for start recording
-  Set,
-
-  /// Currently recording
-  Recording,
-
-  /// This specific recording Stopped, cannot be start again
-  Stopped,
-}
-
-class _RecordButtonState extends State<RecordButton> {
-  IconData _recordIcon = Icons.mic_none;
-  String _recordText = 'Click to start';
-  RecordingState _recordingState = RecordingState.UnSet;
-
-  //Recorder Properties
-  FlutterAudioRecorder audioRecorder;
-
-  @override
-  void initState() {
-    super.initState();
-
-    FlutterAudioRecorder.hasPermissions.then((hasPermision) {
-      if (hasPermision) {
-        _recordingState = RecordingState.Set;
-        _recordIcon = Icons.mic;
-        _recordText = 'Record';
-      }
-    });
-  }
-
-  @override
-  void dispose() {
-    _recordingState = RecordingState.UnSet;
-    audioRecorder = null;
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Stack(
-      alignment: Alignment.center,
-      children: [
-        SizedBox(
-          width: 100,
-          child: RaisedButton(
-            onPressed: () async {
-              await _onRecordButtonPressed();
-              setState(() {});
-            },
-            shape: CircleBorder(
-                //borderRadius: BorderRadius.circular(30),
-                ),
-            child: Container(
-              width: 150,
-              height: 150,
-              child: Icon(
-                _recordIcon,
-                size: 50,
-              ),
-            ),
-          ),
-        ),
-        Align(
-          alignment: Alignment.bottomCenter,
-          child: Padding(
-            child: Text(_recordText),
-            padding: const EdgeInsets.all(8),
-          ),
-        )
-      ],
-    );
-  }
-
-  Future<void> _onRecordButtonPressed() async {
-    switch (_recordingState) {
-      case RecordingState.Set:
-        await _recordVoice();
-        break;
-
-      case RecordingState.Recording:
-        await _stopRecording();
-        _recordingState = RecordingState.Stopped;
-        _recordIcon = Icons.fiber_manual_record;
-        _recordText = 'Record new one';
-        break;
-
-      case RecordingState.Stopped:
-        await _recordVoice();
-        break;
-
-      case RecordingState.UnSet:
-        Scaffold.of(context).hideCurrentSnackBar();
-        Scaffold.of(context).showSnackBar(SnackBar(
-          content: Text('Please allow recording from settings.'),
-        ));
-        break;
-    }
-  }
-
-  Future<void> _recordVoice() async {
-    if (await FlutterAudioRecorder.hasPermissions) {
-      await _initRecorder();
-
-      await _startRecording();
-      _recordingState = RecordingState.Recording;
-      _recordIcon = Icons.stop;
-      _recordText = 'Recording';
-    } else {
-      Scaffold.of(context).hideCurrentSnackBar();
-      Scaffold.of(context).showSnackBar(SnackBar(
-        content: Text('Please allow recording from settings.'),
-      ));
-    }
-  }
-
-  _initRecorder() async {
-    Directory appDirectory = await getApplicationDocumentsDirectory();
-    String filePath = appDirectory.path +
-        '/' +
-        DateTime.now().millisecondsSinceEpoch.toString() +
-        '.aac';
-
-    audioRecorder =
-        FlutterAudioRecorder(filePath, audioFormat: AudioFormat.AAC);
-    await audioRecorder.initialized;
-  }
-
-  _startRecording() async {
-    await audioRecorder.start();
-    // await audioRecorder.current(channel: 0);
-  }
-
-  _stopRecording() async {
-    await audioRecorder.stop();
-
-    widget.onSaved();
-  }
-}
